pyproject.toml
[project]
name = "slangpy-volume-rendering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.0",
    "pygltflib>=1.16.5",
    "ruff>=0.14.14",
    "slangpy>=0.40.1",
    "tqdm>=4.67.3",
]

[project.scripts]
slangpy-volume-rendering = "slangpy_volume_rendering.__main__:main"

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"